requirements.txt
# Core dependencies
langchain==0.0.350
openai==1.78.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.1.0
streamlit==1.44.1
tabulate==0.9.0
numpy==1.26.4
